feat: change interfaces according API docs in dashboard service (#1)
